ARM: dts: Fix LEDs on exynos5422-odroidxu3
The LEDs on Odroid XU3 family boards could not properly probe because PWM
outputs were reduced only to PWM for fan. Fix it for Odroid XU3 and
XU3-Lite boards by switching to usage of all 4 outputs (although the PWM
for MIPI probably is redundant because board does not have MIPI
connector available).

This fixes warnings on dmesg:
[    4.838712] samsung-pwm 12dd0000.pwm: tried to request PWM channel 1 without output
[    4.838725] leds_pwm pwmleds: unable to request PWM for green:mmc0: -22
[    4.838767] leds_pwm: probe of pwmleds failed with error -22

Signed-off-by: Krzysztof Kozlowski <k.kozlowski.k@gmail.com>