x86/mce: Fix thermal throttling reporting after kexec
The per CPU thermal vector init code checks if the thermal vector is
already installed and complains and bails out if it is.

This happens after kexec, as kernel shut down does not clear the thermal
vector APIC register.

This causes two problems:

1. So we always do not fully initialize thermal reports after kexec. The
CPU is still likely initialized, as the previous kernel should have done
it. But we don't set up the software pointer to the thermal vector, so
reporting may end up with a unknown thermal interrupt message.

2. Also it complains for every logical CPU, even though the value is
actually derived from BP only.

The problem is that we end up with one message per CPU, so on larger
systems it becomes very noisy and messes up the otherwise nicely
formatted CPU bootup numbers in the kernel log.

Just remove the check. I checked the code and there's no valid code
paths where the thermal init code for a CPU could be called multiple
times.

Why the kernel does not clean up this value on shutdown:

The thermal monitoring is controlled per logical CPU thread. Normal
shutdown code is just running on one CPU. To disable it we would need a
broadcast NMI to all CPUs on shut down. That's overkill for this. So we
just ignore it after kexec.

Signed-off-by: Andi Kleen <ak@linux.intel.com>
Reviewed-by: Thomas Gleixner <tglx@linutronix.de>
Cc: "H. Peter Anvin" <hpa@zytor.com>
Cc: Ingo Molnar <mingo@redhat.com>
Cc: Tony Luck <tony.luck@intel.com>
Cc: linux-edac <linux-edac@vger.kernel.org>
Cc: x86-ml <x86@kernel.org>
Link: http://lkml.kernel.org/r/1444681922-8644-1-git-send-email-andi@firstfloor.org
Signed-off-by: Borislav Petkov <bp@suse.de>