PM / OPP: Remove confusing error message in of_cpumask_init_opp_table()
The of_cpumask_init_opp_table() function will print an error message
(with pr_err) if it cannot find the OPP table for a certain CPU in the
Device Tree.

There are users of the cpufreq-dt driver (which is the one calling
of_cpumask_init_opp_table) that do not have the OPP points defined in
the Device Tree. Instead, such users dynamically create the OPP table
at boot time depending on the system configuration. Such a case is
planned in the cpufreq-dt driver, which on purpose ignores the return
value of of_cpumask_init_opp_table() with the following comment: "OPPs
might be populated at runtime, don't check for error here".

For such platforms, the of_cpumask_init_opp_table() prints a spurious
and confusing error message for each CPU:

[    1.749548] of_cpumask_init_opp_table: couldn't find opp table for cpu:0, -19
[    1.756784] of_cpumask_init_opp_table: couldn't find opp table for cpu:1, -19
[    1.764031] of_cpumask_init_opp_table: couldn't find opp table for cpu:2, -19
[    1.771268] of_cpumask_init_opp_table: couldn't find opp table for cpu:3, -19

This is confusing because everything is working fine, cpufreq works
and it knows the OPP table that was registered at boot time (on
Marvell Armada XP):

$ cat /sys/devices/system/cpu/cpu0/cpufreq/stats/time_in_state
666500 1884
1333000 23333

To avoid this confusion, this patch simply deletes the error message.

Fixes: 2e02d8723edf cpufreq: dt: Add support for operating-points-v2 bindings
Signed-off-by: Thomas Petazzoni <thomas.petazzoni@free-electrons.com>
Signed-off-by: Rafael J. Wysocki <rafael.j.wysocki@intel.com>