block, dax: opt-in control for raw block dax support
Now that we have the ability to dynamically enable/disable DAX for a raw
block inode, make the default behavior a compile time decision.  DAX
does not yet have feature parity with pagecache backed mappings, and it
may disable statistics that an application depends on, so environments
should knowingly enable DAX semantics.

Note, that this does not affect the mmap path for filesystems on top of
a DAX capable block device.  They currently open code a check for the
->direct_access() op in the gendisk.  That said, DAX support is already
opt-in for filesystems via a mount flag.

Cc: Dave Chinner <david@fromorbit.com>
[dgc: leave the dax_do_io() path alone, let it honor S_DAX]
Signed-off-by: Dan Williams <dan.j.williams@intel.com>