cpufreq: intel_pstate: Don't use ACPI perf for HWP
When HWP is enabled, the information in _PSS and _PPC prevents entering
turbo states. The reason is HWP and legacy pstate are not compatible.
There is no processing of turbo activation ratio in HWP, but _PPC setting
still behaves as there is support. So we end up in a pstate, which is
below or equal to max non turbo ratio.

Also the max turbo ratio is set to wrong value in _PSS, so even if HWP
is disabled to use acpi-cpufrq or intel_pstate in legacy mode, this will
still cause issues.  This change will not read any _PSS table or call
acpi_processor_register_performance, when the platform has HWP feature.

Fixes: 37afb0003242 (cpufreq: intel_pstate: Use ACPI perf configuration)
Signed-off-by: Srinivas Pandruvada <srinivas.pandruvada@linux.intel.com>
[ rjw: Rebase ]
Signed-off-by: Rafael J. Wysocki <rafael.j.wysocki@intel.com>