ARM: 8452/1: PJ4: move coprocessor register access sequences to iwmmxt.S
The PJ4 inline asm sequences in pj4-cp0.c cannot be built in Thumb-2 mode,
due to the way it performs arithmetic on the program counter, so it is
built in ARM mode instead. However, building C files in ARM mode under
CONFIG_THUMB2_KERNEL is problematic, since the instrumentation performed
by subsystems like ftrace does not expect having to deal with interworking
branches.

So instead, revert to building pj4-cp0.c in Thumb-2 mode, and move the
offending sequence to iwmmxt.S, which is not instrumented anyway, and is
already built in ARM mode unconditionally.

Fixes: 13d1b9575ac2 ("ARM: 8221/1: PJ4: allow building in Thumb-2 mode")
Reported-by: Stephen Boyd <sboyd@codeaurora.org>
Tested-by: Stephen Boyd <sboyd@codeaurora.org>
Signed-off-by: Ard Biesheuvel <ard.biesheuvel@linaro.org>
Signed-off-by: Russell King <rmk+kernel@arm.linux.org.uk>