mm, debug: fix wrongly filtered flags in dump_vma()
This is the second version of patchset which originally aimed to improve the
page_owner functionality.

For page_owner, the main changes are

o Use static key to further reduce overhead when compiled in but not enabled.
o Improve output wrt. page and pageblock migratetypes
o Transfer the info on page migrations and track last migration reason.
o Dump the info as part of dump_page() to hopefully help debugging.

For the last point, Kirill requested a human readable printing of gfp_mask
and migratetype after v1.  At that point it probably makes a lot of sense
to do the same for page alloc failure and OOM warnings.  The flags have
been undergoing revisions recently, and we might be getting reports from
various kernel versions that differ.  The ./scripts/gfp-translate tool
needs to be pointed at the corresponding sources to be accurate.  The
downside is potentially breaking scripts that grep these warnings, but
it's not a first change done there over the years.

Note I'm not entirely happy about the dump_gfpflag_names() implementation,
due to usage of pr_cont() unreliable on SMP (and I've seen spurious
newlines in dmesg output, while being correct on serial console or
/var/log/messages).  It also doesn't allow plugging the gfp_mask
translation into /sys/kernel/debug/page_owner where it also could make
sense.  Maybe a new *printf formatting flag?  Too specialized maybe?  Or
just prepare the string in a buffer on stack with strscpy?

This patch (of 9):

The dump_vma() function uses dump_flags() for printing the flags as
symbolic names.  That function however does a page-flags specific
filtering of bits higher than NR_PAGEFLAGS in order to remove the zone id
part.  For dump_vma() this results in removing several VM_* flags from the
symbolic translation.

Fix this by refactoring dump_flags() to dump_flag_names(), which only
prints the symbolic names in parentheses.  Printing the raw flag value
with a prefix, and any filtering is left to the caller.  In addition to
fixing the bug, this allows better flexibility, which will be useful to
print gfp_flags by a later patch.

Signed-off-by: Vlastimil Babka <vbabka@suse.cz>
Cc: Joonsoo Kim <iamjoonsoo.kim@lge.com>
Cc: Minchan Kim <minchan@kernel.org>
Cc: Sasha Levin <sasha.levin@oracle.com>
Cc: "Kirill A. Shutemov" <kirill.shutemov@linux.intel.com>
Cc: Mel Gorman <mgorman@suse.de>
Cc: Michal Hocko <mhocko@suse.cz>
Signed-off-by: Andrew Morton <akpm@linux-foundation.org>