ARM: qcom: Make an option for qcom 'a-family' platforms
We want to remove the ARCH_MSM* configs in mach-qcom/Kconfig
because they are proxy configs for selecting the right
clocksource driver and adjusting the text offset and there is no
difference between ARCH_MSM8X60 and ARCH_MSM8960 anymore.
Furthermore, there aren't configs for all the SoCs supported by
the mainline kernel, leading to confusion about which SoCs are
supported.

So, add a new config that covers all the platforms that use the
qcom clocksource driver and need the 2MB text offset adjustment
and make it def_bool y so that the defconfigs keep working
properly.

Cc: Arnd Bergmann <arnd@arndb.de>
Signed-off-by: Stephen Boyd <sboyd@codeaurora.org>