coresight: tmc: making disable function reusable
When disabling a TMC in any mode it may not be automatically
desirable to read the content of the trace buffer.  It is
the case when operating in Perf mode where the content of the
trace buffer is dealt with prior to the TMC being disabled.

As such moving the tmc_etb/etr_dump() functions out of the
disable_hw callback and into the current sysFS interface.

Signed-off-by: Mathieu Poirier <mathieu.poirier@linaro.org>