memcg: ignore partial THP when moving task
After "mm: rework mapcount accounting to enable 4k mapping of THPs" it is
possible to have a partial THP accessible via ptes.  Memcg task migration
code is not prepared for this situation and uncharges the tail page from
the original memcg while the original THP is still charged via the head
page which is not mapped to the moved task.  The page counter of the
origin memcg will underflow when the whole THP is uncharged later on and
lead to: WARNING: CPU: 0 PID: 1340 at mm/page_counter.c:26
page_counter_cancel+0x34/0x40() reported by Minchan Kim.

This patch prevents from the underflow by skipping any partial THP pages
in mem_cgroup_move_charge_pte_range.  PageTransCompound is checked when we
do pte walk.  This means that a process might leave a partial THP behind
in the original memcg if there is no other process mapping it via pmd but
this is considered acceptable because it shouldn't happen often and this
is not considered a memory leak because the original THP is still
accessible and reclaimable.  Moreover the task migration has always been
racy and never guaranteed to move all pages.

Reported-by: Minchan Kim <minchan@kernel.org>
Acked-by: Johannes Weiner <hannes@cmpxchg.org>
Signed-off-by: Michal Hocko <mhocko@suse.com>
Cc: "Kirill A. Shutemov" <kirill.shutemov@linux.intel.com>
Signed-off-by: Andrew Morton <akpm@linux-foundation.org>