PCI: rcar: Support runtime PM link state L1 handling in pcie-rcar
The R-Car PCIe host controller does not handle L1 ASPM.  Instead, the
hardware needs assistance to transition to L1.  When the controller has
received a PM_ENTER_L1 DLLP, we can't access a card's config regs until we
have got it out of L1 link state.  The host controller will handle this as
long as it has also been transitioned to L1 link state.

So, when attempting a config access, check to see if the card has gone into
L1, and if so, do the same for the host controller.

This is based on a patch by Hien Dang <hien.dang.eb@rvc.renesas.com>

Signed-off-by: Phil Edworthy <phil.edworthy@renesas.com>
Signed-off-by: Bjorn Helgaas <bhelgaas@google.com>
Acked-by: Wolfram Sang <wsa+renesas@sang-engineering.com>