block: clarify badblocks lifetime
The badblocks list attached to a gendisk is optionally allocated by the
driver which implies the driver owns the lifetime of the object.  Do not
automatically free it in del_gendisk(), leave that to the driver.  This
is in preparation for expanding the use of badblocks in libnvdimm
drivers and introducing devm_alloc_badblocks().

Signed-off-by: Dan Williams <dan.j.williams@intel.com>